Support int8 migration versions via new int8-versions feature (#330)
* Support int8 migration versions via new int8-versions feature

This addresses, though does not really *fix* #83, because it doesn't make refinery support timestamped migrations *by default*, but only if you opt-in to the new feature.
However, making it an optional feature neatly sidesteps the unanswered questions in the issue, and so makes the implementation easier to complete and land.

* cargo clippy

* update wording

* remove repr

* fix enum migrations

* enable int8-versions test

* fix PR

---------

Co-authored-by: João Oliveira <hello@jxs.pt>